requirements.txt
PyYAML>=6.0
aiohttp>=3.9
matplotlib>=3.7,<4.0
numpy>=1.24,<2.0
pandas>=1.5,<3.0
requests>=2.31,<3.0
streamlit>=1.31,<2.0
websockets>=11.0,<12.0

# Optional dependencies
python-binance>=1.0.19,<2.0  # Enables live trading via the official Binance client
mplfinance>=0.12.10b0  # Advanced candlestick rendering
plotly>=5.18,<6.0  # Interactive charts and Dash integration
bokeh>=3.3,<4.0  # Browser-based interactive charts
dash>=2.14,<3.0  # Local web dashboard for Plotly figures
